% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalFunc.R
\name{MD_pvalue}
\alias{MD_pvalue}
\title{MD_pvalue}
\usage{
MD_pvalue(p.r, p.o, N.r, N)
}
\arguments{
\item{p.r}{A vector for the relative abundance of corresponding
taxa in group 1 (e.g. the microbial relative abundance in rectum).}

\item{p.o}{A vector for the relative abundance of corresponding
taxa in group 2 (e.g. the microbial relative abundance in other GI locations).}

\item{N.r}{total counts across all samples in group 1 (i.e. the sum of sequence
depths of all samples in group 1).}

\item{N}{total counts across all samples in group 2 (i.e. the sum of sequence
depths of all samples in group 2).}
}
\value{
A scalar in [0, 1].
}
\description{
This function returns a p-value of testing differential abundance
under the null hypothesis that p.r = p.o.
}
\examples{
p1 <- rep(0.1, 10)
p2 <- c(rep(0.09, 5), rep(0.11, 5))
MD_pvalue(p1, p2, 5000, 5000);
}
