% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicroDemix_EM.R
\name{MicroDemix_EM}
\alias{MicroDemix_EM}
\title{MicroDemix_EM}
\usage{
MicroDemix_EM(data1, data2, cova, maxiter = 10)
}
\arguments{
\item{data1}{stool (mixture) microbiome data set with taxa in
columns and samples in rows.}

\item{data2}{rectum (reference) microbiome data set with taxa in
columns and samples in rows. The taxa in data1 and data2 should
be matching.}

\item{cova}{covariates related to samples in data1 that have a matching order
with data1. 'cova' must be a matrix with covariates in columns and observations
in rows.}

\item{maxiter}{The maximum number of iterations taken by the EM algorithm. Defaults to 10.}
}
\value{
A list containing the following components:
\item{pr.est}{the estimated relative abundance for the rectum (reference) microbiome}
\item{ps.est}{the estimated relative abundance for the stool (mixture) microbiome}
\item{po.est}{the estimated relative abundance for the microbiome in other GI locations}
}
\description{
This function using the EM algorithm to estimate the microbial relative abundance in stool (mixture),
rectum (reference) and other GI locations.
}
\examples{
MicroDemix_EM(yig, yig.n, x, maxiter=5);
}
