% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicroDemix_EM_permutation.R
\name{MicroDemix_EM_permutation}
\alias{MicroDemix_EM_permutation}
\title{MicroDemix_EM_permutation}
\usage{
MicroDemix_EM_permutation(data1, data2, cova1, cova2, maxiter = 10, B)
}
\arguments{
\item{data1}{stool (mixture) microbiome data set with taxa in
columns and samples in rows.}

\item{data2}{rectum (reference) microbiome data set with taxa in
columns and samples in rows. The taxa in data1 and data2 should
be matching.}

\item{cova1}{covariates related to samples in data1 that have a matching order
with data1. 'cova1' must be a matrix with covariates in columns and observations
in rows.}

\item{cova2}{covariates related to samples in data2 that have a matching order
with data2. 'cova2' must be a matrix with covariates in columns and observations
in rows.}

\item{maxiter}{The maximum number of iterations taken by the EM algorithm. Defaults to 10.}

\item{B}{The number of permutations.}
}
\value{
The p-value. A scalar in [0, 1].
}
\description{
This function performs a permutation based test of differential abundance
under the null hypothesis that p.r = p.o. Variability in the results is anticipated due to the
randomness embedded in this function.
}
\examples{
MicroDemix_EM_permutation(yig, yig.n, x, x.n, maxiter=5, B=10);
}
