% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimationFunc.R
\name{RA_plot}
\alias{RA_plot}
\title{RA_plot}
\usage{
RA_plot(p.r, p.s, p.o, G, taxon)
}
\arguments{
\item{p.r}{relative abundance at location 1 (e.g. rectum)}

\item{p.s}{relative abundance at location 2 (e.g. stool)}

\item{p.o}{relative abundance at location 3 (e.g. other GI locations).
The taxa in p.r, p.s, and p.o should be matching.}

\item{G}{the number of taxa in p.r}

\item{taxon}{the names of taxa in p.r}
}
\value{
A stack plot
}
\description{
visualization of microbial relative abundance (proportions) at three different locations.
}
\examples{
p1 <- rep(0.2, 5)
p2 <- c(0.1, 0.2, 0.25, 0.25, 0.2)
p3 <- c(rep(0.1, 3),rep(0.35, 2))
RA_plot(p1,p2,p3,5,c("A","B","C","D","E"));
}
