% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x.n}
\alias{x.n}
\title{A sample covariates data related to samples in data 'yig.n'}
\format{
An object of "matrix" "array".
}
\usage{
data(x.n)
}
\description{
A simulated covariates data corresponding to samples in data 'yig.n'.
}
\examples{
data(x.n)
dim(x.n)
}
\keyword{datasets}
