% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x}
\alias{x}
\title{A sample covariates data related to samples in data 'yig'}
\format{
An object of "data.frame".
}
\usage{
data(x)
}
\description{
Data set contains information on consent_age and gender corresponding to
samples in data 'yig'.
}
\examples{
data(x)
colnames(x)
}
\keyword{datasets}
