% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yig.n}
\alias{yig.n}
\title{A sample rectum (reference) microbiome data 'yig.n'}
\format{
An object of "matrix" "array".
}
\usage{
data(yig.n)
}
\description{
A sample simulated microbiome data with 10 taxa in column and 100 samples in row.
}
\examples{
data(yig.n)
dim(yig.n)
}
\keyword{datasets}
