% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yig}
\alias{yig}
\title{A sample stool (mixture) microbiome data 'yig'}
\format{
An object of "matrix" "array".
}
\usage{
data(yig)
}
\description{
A sample simulated microbiome data with 10 taxa in column and 100 samples in row.
}
\examples{
data(yig)
dim(yig)
}
\keyword{datasets}
