% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimationFunc.R
\name{MD_estimate}
\alias{MD_estimate}
\title{MD_estimate}
\usage{
MD_estimate(data1, data2, cova)
}
\arguments{
\item{data1}{stool (mixture) microbiome data set with taxa in
column and samples in row.}

\item{data2}{rectum (reference) microbiome data set with taxa in
column and samples in row. The taxa in data1 and data2 should
be matching.}

\item{cova}{covariates related to samples in data1 that have a matching order
with data1.}
}
\value{
A list containing the following components:
\item{pr.est}{the estimated relative abundance for the rectum (reference) microbiome}
\item{ps.est}{the estimated relative abundance for the stool (mixture) microbiome}
\item{po.est}{the estimated relative abundance for the microbiome in other GI locations}
}
\description{
This function estimate the microbial relative abundance in stool (mixture),
rectum (reference) and other GI locations.
}
\examples{
MD_estimate(yig, yig.n, x);
}
