% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yig.n}
\alias{yig.n}
\title{A sample rectum (reference) microbiome data 'yig.n'}
\format{
An object of "matrix" "array".
}
\usage{
data(yig.n)
}
\description{
Microbiome data with 9 families in column and 30 samples in row.
}
\examples{
data(yig.n)
colnames(yig.n)
}
\keyword{datasets}
