% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yig}
\alias{yig}
\title{A sample stool (mixture) microbiome data 'yig'}
\format{
An object of "matrix" "array".
}
\usage{
data(yig)
}
\description{
Microbiome data with 9 families in column and 12 samples in row.
}
\examples{
data(yig)
colnames(yig)
}
\keyword{datasets}
